Architectural refactor
