+++ src/main/java/pickleib/web/steps/WebStepBase.java
@@ -3,7 +3,7 @@ package pickleib.web.steps;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.openqa.selenium.remote.RemoteWebDriver;
 import pickleib.utilities.ElementAcquisition;
-import pickleib.utilities.ScreenCaptureUtility;
+import pickleib.utilities.screenshot.ScreenCaptureUtility;
 import pickleib.web.driver.PickleibWebDriver;
 import pickleib.web.interactions.WebInteractions;
 import utils.Printer;
